User Registration in DB added

+++ app/src/main/java/com/example/touristaapp/models/User.java
@@ -109,7 +109,7 @@ public class User {
         this.password = password;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
